# HEX-LoadBench/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HEX-LoadBench API
  description: |
    Load-Testing & API Stress Test Platform
    
    ## WARNING
    THIS TOOL IS FOR AUTHORIZED PERFORMANCE TESTING ONLY.
    
    Users are responsible for ensuring they have proper authorization before testing any systems.
    
    ## Authentication
    This API uses JWT tokens and API keys for authentication.
    
    - JWT tokens are used for user authentication (session-based)
    - API keys are used for service-to-service communication
    - All endpoints require authentication except `/health` and `/docs`
    
  version: 1.0.0
  contact:
    name: HEX-LoadBench Support
    email: support@hexloadbench.com
    url: https://hexloadbench.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hexloadbench.com
    description: Production server
paths:
...
  /users/me:
    put:
      tags:
        - Users
      summary: Update current user profile
      description: Update profile information for the current user
      operationId: updateCurrentUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/api-keys:
    get:
      tags:
        - Users
      summary: List API keys
      description: Returns all API keys belonging to the current user
      operationId: listApiKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API keys retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: number
          format: float
          description: Unix timestamp
...
    UserRole:
      type: string
      enum:
        - admin
        - operator
        - viewer
      description: User role in the system
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        email:
          type: string
          format: email
          description: User email
        username:
          type: string
          description: Unique username
        full_name:
          type: string
          description: User's full name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          description: Whether the user is active
        is_verified:
          type: boolean
          description: Whether the user email is verified
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        last_login:
          type: string
          format: date-time
          description: Last login timestamp
        organization_id:
          type: integer
          description: Organization ID
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: New email address
        username:
          type: string
          description: New username
        full_name:
          type: string
          description: New full name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          description: Account active status
...
    APIKeyResponse:
      type: object
      properties:
        key_id:
          type: string
          description: API key identifier
        name:
          type: string
          description: API key name
        description:
          type: string
          description: API key description
        scopes:
          type: array
          items:
            type: string
          description: List of allowed scopes
        is_active:
          type: boolean
          description: Whether the API key is active
        expires_at:
          type: string
          format: date-time
          description: API key expiration date
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        last_used:
          type: string
          format: date-time
          description: Last usage timestamp
        usage_count:
          type: integer
          description: Number of times the key was used
...